knowledge-search-api/pyproject.toml
[tool.poetry]
name = "knowledge-search-api"
version = "0.1.0"
description = "A search microservice providing RAG and Knowledge Graph capabilities."
authors = ["Your Name <you@example.com>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"  # <-- ИСПРАВЛЕНО
fastapi = "^0.111.0"
uvicorn = {extras = ["standard"], version = "^0.29.0"}
psycopg2-binary = "^2.9.9"
neo4j = "^5.19.0"
sentence-transformers = "^2.7.0"
python-dotenv = "^1.0.1"
tiktoken = "^0.7.0"
requests = "^2.31.0"
torch = {version = "^2.6.0", source = "pytorch_cu124"}
torchvision = {version = "^0.21.0", source = "pytorch_cu124"}
torchaudio = {version = "^2.6.0", source = "pytorch_cu124"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[[tool.poetry.source]]
name = "pytorch_cu124"
url = "https://download.pytorch.org/whl/cu124"
priority = "explicit"